requirements.txt
python-dotenv
cx_Oracle
pytest